Add files via upload
Here is the final version of part III question 6. Which is that different version of the perceptron. I made a One-vs-One classifier rather than a One-vs-Rest which is what our previous one was. There are 2 results, part 1 is called "basic_results.csv" and part 2 is saved in the notebook in the "cross-val' section.